fs_task/requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2022.1
Django==4.2.4
django-crispy-forms==2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.2.0
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
